+++ carp_core/lib/client/datapoint.dart
@@ -68,7 +68,7 @@ part of carp_core_client;
 //   DataPointHeader carpHeader;
 
 //   /// The CARP data point body. Can be any payload modelled as a [Data].
-//   @JsonKey(ignore: true)
+//   @JsonKey(includeFromJson: false, includeToJson: false)
 //   Data? data;
 
 //   Map<String, dynamic>? _carpBody;