+++ carp_core/lib/deployment/infrastructure/deployment_requests.dart
@@ -42,7 +42,7 @@ class CreateStudyDeployment extends DeploymentServiceRequest {
   StudyProtocol protocol;
 
   @override
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   // ignore: overridden_fields
   String? studyDeploymentId;