+++ carp_mobile_sensing/lib/sampling_packages/device/device.dart
@@ -17,6 +17,7 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:screen_state/screen_state.dart';
 import 'package:system_info2/system_info2.dart';
+import 'package:flutter_timezone/flutter_timezone.dart';
 
 part 'device.g.dart';
 part 'device_data.dart';